requirements.txt
opencv-python
flask
mysql-connector-python